requirements.txt
flask>=2.2
langchain>=0.0.300
google-cloud-aiplatform>=1.28.0
PyPDF2>=3.0.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
uvicorn>=0.21.1
gunicorn>=20.1.0
sentence-transformers>=2.2.2